+++ changelog.md
@@ -2,9 +2,13 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9456]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9478]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9478] - 2025-12-13
+Device Tsp
+- Condition getting script info state on previous scripts.
 
 ## [8.1.9457] - 2025-11-22
 Device Tsp